backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
websockets==14.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Agent framework
langgraph==0.2.60
langchain==0.3.12
langchain-google-genai==2.0.8
langchain-community==0.3.12
langchain-chroma==0.1.4

# Google Gemini
google-generativeai==0.8.3

# Vector store
chromadb==0.5.23

# Observability
langfuse==2.57.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# HTTP client
httpx==0.28.1
aiohttp==3.11.11

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
Pillow==11.1.0
gitpython==3.1.43

# Security
bleach==6.2.0

# Email
resend==2.7.0
